requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
python-dotenv
cryptography
pydantic
